Type cast a negative uint64_t to int64_t to avoid confusion when doing arithmetic with it in accounting dealing with over commit time.

+++ branches/slurm-2.0/src/plugins/accounting_storage/mysql/mysql_rollup.c
@@ -804,8 +804,8 @@ extern int mysql_hourly_rollup(mysql_conn_t *mysql_conn,
 			if((int64_t)c_usage->i_cpu < 0) {
 /* 				info("got %d %d %d", c_usage->r_cpu, */
 /* 				     c_usage->i_cpu, c_usage->o_cpu); */
-				c_usage->r_cpu += c_usage->i_cpu;
-				c_usage->o_cpu -= c_usage->i_cpu;
+				c_usage->r_cpu += (int64_t)c_usage->i_cpu;
+				c_usage->o_cpu -= (int64_t)c_usage->i_cpu;
 				c_usage->i_cpu = 0;
 				if((int64_t)c_usage->r_cpu < 0)
 					c_usage->r_cpu = 0;
